traffic_congestion/pyproject.toml
[tool.poetry]
name = "traffic-congestion"
version = "0.1.0"
description = "Using AI to support the local community problem."
authors = ["tqtensor - <victortang212@gmail.com>"]
readme = "README.md"
packages = [{include = "traffic_congestion"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
ffmpy = "^0.3.0"
google-cloud-storage = "^2.6.0"
m3u8 = "^3.3.0"
py = "^1.11.0"
pydub = "^0.25.1"
pytest = "^7.2.0"
requests = "^2.28.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
yapf = "^0.32.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
